Native Image Support
This change adds Native Image support for Spring Boot applications.  It does
this by requesting GraalVM and the native-image utility be installed.  Then it
takes the compiled Java applications and runs the native-image utility against
it.  Finally, it removes the original application and creates process types
that point at the compiled native image.

Signed-off-by: Ben Hale <bhale@vmware.com>