JAR Root on Classpath
Previously, the native-image classpath did not include the JAR's root.
Since the native image would run without needed the Spring Boot
launcher code, I thought this would be sufficient.  However, there are
other files, specifically in META-INF/ that are useful to native-image
and therefore this directory does need to be on the classpath.  This
change adds it.

Signed-off-by: Ben Hale <bhale@vmware.com>