Version Warnings
This change adds support for warning based on the support status of a given
Spring Boot generation.  At this time, we output a message if open source or
commercial support has ended for a given application's Spring Boot version.
The machinery is in place to start warning about versions of other
dependencies as well if we think there is enough value in that.

[resolves #4]

Signed-off-by: Ben Hale <bhale@vmware.com>