Adjust BPL_JVM_THREAD_COUNT for Reactive Web Applications
Previously all applications, including reactive web applications, used the
default thread count (250) during memory calculation.  This was wasted memory
for reactive web applications which have one thread per logical execution
core.  This change adds support for detecting that an application is a
reactive web application and if it is, overrides the default thread count,
setting it to 50 if the user has not already specified a value.

[resolves #7]

Signed-off-by: Ben Hale <bhale@vmware.com>