New Data Flow Properties Name
Previously the name of the Data Flow properties file wasn't as inclusive as it
could have been.  The team has now deprecated the old name and introduced a
new, more inclusive name.  This change updates the buildpack to detect and
prefer this new name, while falling back to the deprecated name if it does not
exist.  This is a temporary state before the deprecated name is removed
completely from both Data Flow and the buildpack.

[resolves #13][spring-cloud/spring-cloud-dataflow#4053]

Signed-off-by: Ben Hale <bhale@vmware.com>