Contributes exanded classpath during native image builds
Contribute appends application classes and entries from classpath.idx to the build time classpath.

In a JVM application the Spring Boot Loader would make these modifications to the classpath when the executable JAR
or WAR is launched. To set the classpath properly for a native-image build the buildpack must replicate this behavior.

Signed-off-by: Emily Casey <ecasey@vmware.com>